refactor: adopt feature based architecture for maintanability & resusability
