+++ src/features/experience/components/ExperienceCard.astro
@@ -1,0 +1,108 @@
+---
+import { Image } from 'astro:assets';
+import { Icon } from 'astro-icon/components';
+
+import Button from '@shared/components/base/Button.astro';
+import type { Experience } from '@features/experience/types';
+
+type Props = Experience;
+const {
+	iconName,
+	company,
+	title,
+	type,
+	location,
+	date,
+	responsibilities,
+	image,
+	showCV,
+	companyURL,
+} = Astro.props;
+---
+
+<article
+	class="card-hover group/exp-card relative z-10 flex flex-col gap-4 rounded-xl border border-zinc-200 bg-zinc-50 py-4 pl-4 pr-8 dark:border-zinc-800 dark:bg-zinc-900 lg:py-8 lg:pl-8 lg:pr-12"
+>
+	<a href={companyURL} target="_blank" class="link-focus flex w-max flex-row items-center gap-4">
+		<div>
+			{
+				(
+					<Icon
+						name={iconName}
+						width={iconName === 'bangor-university' ? 65 : 48}
+						height={iconName === 'bangor-university' ? 65 : 48}
+						class={iconName === 'bangor-university' ? 'rounded bg-white p-px' : ''}
+					/>
+				)
+			}
+		</div>
+		<span class="text-xl text-zinc-700 dark:text-zinc-300 xl:text-2xl">
+			{company}
+		</span>
+	</a>
+	<div class="flex flex-col gap-8">
+		<div class="flex flex-col gap-2">
+			<div class="flex flex-row items-center gap-2">
+				<h2
+					class="text-3xl font-medium text-zinc-700 group-hover/exp-card:text-zinc-900 dark:text-zinc-300 dark:group-hover/exp-card:text-zinc-100 xl:text-4xl"
+				>
+					{title}
+				</h2>
+				<div
+					class="inline-flex items-center whitespace-nowrap rounded-md border border-transparent bg-zinc-900 px-2.5 py-0.5 text-xs font-semibold text-zinc-50 shadow dark:bg-zinc-100 dark:text-zinc-900"
+				>
+					{type}
+				</div>
+			</div>
+			<span class="text-sm font-medium text-zinc-500 dark:text-zinc-400">
+				{location} | <time>{date}</time>
+			</span>
+			<div class="h-px w-full bg-zinc-200 dark:bg-zinc-800"></div>
+			<ul
+				class="list-outside list-disc space-y-4 p-4 text-sm text-zinc-600 group-hover/exp-card:text-zinc-800 dark:text-zinc-300 dark:group-hover/exp-card:text-zinc-200 md:text-base xl:text-lg"
+			>
+				{responsibilities.map((responsibility) => <li>{responsibility}</li>)}
+			</ul>
+			{
+				image ? (
+					<div class="flex max-h-[250px] w-full max-w-[400px] justify-start">
+						<Image
+							src={image}
+							alt={`${company} highlight`}
+							width={925}
+							height={731}
+							loading="lazy"
+							decoding="async"
+							quality={100}
+							class="rounded-md object-cover object-top shadow-xl"
+						/>
+					</div>
+				) : null
+			}
+			{
+				showCV ? (
+					<div class="flex flex-row flex-wrap items-center gap-4">
+						<Button
+							as="a"
+							href="/kieranroberts-CV.pdf"
+							target="_blank"
+							download="kieranroberts-CV.pdf"
+						>
+							Download CV
+							<Icon
+								name="download-circle"
+								width={20}
+								height={20}
+								class="stroke-current transition-transform duration-200 group-hover:translate-y-1 group-hover:text-accent-brightest dark:group-hover:text-accent-darkest"
+							/>
+						</Button>
+						<span class="text-sm text-zinc-500 dark:text-zinc-400">
+							{' '}
+							See my CV for more details.{' '}
+						</span>
+					</div>
+				) : null
+			}
+		</div>
+	</div>
+</article>
