+++ src/features/portfolio/components/HeroIllustration.astro
@@ -1,0 +1,107 @@
+---
+import { Icon } from 'astro-icon/components';
+import { Image } from 'astro:assets';
+
+import kieranAtWorkImage from '@images/kieran-at-work.webp';
+import hashnodeTeam2022Image from '@images/hashnode-team-2022.webp';
+import hashnodeTeam2024Image from '@images/hashnode-team-2024.webp';
+---
+
+<script>
+	function setupAnimations() {
+		const observer = new IntersectionObserver(
+			(entries) => {
+				entries.forEach((entry) => {
+					if (entry.isIntersecting) {
+						const animateAttr = entry.target.getAttribute('data-hero-image-animate');
+						if (animateAttr) {
+							entry.target.classList.add(...animateAttr.split(' '));
+						}
+						observer.unobserve(entry.target);
+					}
+				});
+			},
+			{
+				rootMargin: '0px',
+				threshold: 0,
+			},
+		);
+
+		document.querySelectorAll('[data-hero-image-animate]').forEach((element) => {
+			observer.observe(element);
+		});
+	}
+
+	setupAnimations();
+</script>
+
+<div class="flex flex-col items-center justify-center gap-4">
+	<p class="relative text-center text-sm text-zinc-600 dark:text-zinc-400 lg:text-lg">
+		I love to collaborate
+	</p>
+	<div class="mx-auto h-px w-40 rounded-full bg-zinc-200 dark:bg-zinc-800"></div>
+</div>
+
+<div
+	data-hero-image-animate="opacity-100"
+	class="pointer-events-none absolute left-[58%] top-[78%] z-50 hidden -translate-x-1/2 -translate-y-1/2 opacity-0 transition-opacity delay-300 duration-300 xl:block"
+>
+	<div class="rotate-[180deg] transform">
+		<Icon name="styleised-arrow" width={256} height={256} />
+	</div>
+	<span class="absolute -right-8 bottom-14 inline-block text-base text-zinc-600 dark:text-zinc-400"
+		>Me</span
+	>
+</div>
+
+<div
+	class="relative z-10 mt-12 flex w-full flex-col items-center justify-center gap-4 lg:flex-row lg:gap-0 [&>div:has(~div:hover)]:scale-90 [&>div:has(~div:hover)]:brightness-50 [&>div:hover~div]:scale-90 [&>div:hover~div]:brightness-50"
+>
+	<div
+		data-hero-image-animate="opacity-100"
+		class="z-20 h-[290px] w-full rounded-xl object-cover opacity-0 shadow-xl brightness-[0.8] transition-all duration-150 ease-in hover:brightness-100 lg:max-w-[550px] lg:translate-x-20 lg:rotate-6 lg:hover:z-50 lg:hover:rotate-0 lg:hover:scale-105 xl:max-w-[650px]"
+	>
+		<Image
+			src={hashnodeTeam2022Image}
+			alt="Hashnode Team from 2022 posing for a team photo"
+			class="h-full w-full rounded-xl object-cover object-center"
+			loading="lazy"
+			decoding="async"
+			quality={100}
+			width={1000}
+			height={674}
+		/>
+	</div>
+
+	<div
+		data-hero-image-animate="opacity-100"
+		class="z-30 h-[290px] w-full transform rounded-xl object-cover opacity-0 shadow-xl brightness-[0.8] transition-all duration-150 ease-in lg:max-w-[550px] lg:-rotate-[5deg] lg:hover:z-30 lg:hover:-rotate-0 lg:hover:scale-105 lg:hover:brightness-100 xl:max-w-[650px]"
+	>
+		<Image
+			src={kieranAtWorkImage}
+			alt="Kieran discussing a project with a coworker in person while sharing his laptop screen"
+			class="h-full w-full rounded-xl object-cover"
+			loading="lazy"
+			decoding="async"
+			quality={100}
+			width={1000}
+			height={674}
+		/>
+	</div>
+
+	<div
+		data-hero-image-animate="opacity-100"
+		class="pacity-0 z-20 h-[290px] w-full rounded-xl object-cover opacity-0 shadow-xl brightness-[0.8] transition-all duration-150 ease-in hover:brightness-100 lg:max-w-[550px] lg:-translate-x-20 lg:rotate-6 lg:hover:z-50 lg:hover:rotate-0 lg:hover:scale-105 xl:max-w-[650px]"
+	>
+		<Image
+			src={hashnodeTeam2024Image}
+			alt="Kieran discussing a project with a coworker in person while sharing his laptop screen"
+			class="h-full w-full rounded-xl object-cover"
+			loading="lazy"
+			decoding="async"
+			quality={100}
+			width={1000}
+			height={674}
+		/>
+	</div>
+</div>
