+++ src/features/projects/types/index.ts
@@ -1,0 +1,10 @@
+export type Project = {
+	title: string;
+	description: string;
+	date: string;
+	githubLink: string;
+	liveLink: string | null;
+	caseStudyLink: string | null;
+	image: ImageMetadata;
+	status: 'complete' | 'in progress' | 'planned';
+};
