--- src/layouts/Footer.astro
@@ -1,23 +1,0 @@
----
-import { DOMAIN_NAME } from '@consts/urls';
-import SocialLinkGroup from '@components/SocialLinkGroup.astro';
----
-
-<footer class="border-t dark:border-zinc-800">
-	<div class="mx-auto max-w-[1400px] gap-8 px-4 py-4 md:px-16 lg:px-24 2xl:py-8">
-		<div
-			class="flex w-full flex-col items-center gap-4 text-zinc-600 dark:text-zinc-300 sm:flex-row sm:justify-between"
-		>
-			<div class="flex flex-col gap-1">
-				<a
-					href="/"
-					class="link-focus rounded-md text-lg font-medium hover:text-zinc-900 dark:hover:text-zinc-50"
-				>
-					{DOMAIN_NAME}
-				</a>
-				<div class="text-center text-sm">Design & build by Kieran.</div>
-			</div>
-			<SocialLinkGroup />
-		</div>
-	</div>
-</footer>
