--- src/modules/landing/ProjectGroup.astro
@@ -1,8 +1,0 @@
----
-import ProjectCard from '@modules/landing/ProjectCard.astro';
-import { projects } from '@consts/index';
----
-
-<div class="grid grid-cols-1 gap-4 md:grid-cols-2">
-	{projects.map((project) => <ProjectCard {...project} />)}
-</div>
