--- src/modules/landing/TestimonialCard.astro
@@ -1,51 +1,0 @@
----
-import type { Testimonial } from '@env';
-import { Image } from 'astro:assets';
-
-interface Props extends Testimonial {
-	center: boolean;
-}
-
-const { name, company, role, description, image, profileURL, center } = Astro.props;
----
-
-<article
-	class={`card-hover group z-10 overflow-hidden rounded-xl border border-zinc-200 bg-white dark:border-zinc-800 dark:bg-zinc-950 ${center ? 'md:col-start-2 md:col-end-6' : 'md:col-span-3'}`}
->
-	<div class="flex h-full flex-col justify-between py-8 pl-6 pr-8">
-		<div class="flex flex-1 items-center">
-			<p
-				class="text-center text-base text-zinc-500 transition-[text-color] duration-300 ease-in group-hover:text-zinc-800 dark:text-zinc-400 dark:group-hover:text-zinc-200 lg:text-lg"
-			>
-				"{description}"
-			</p>
-		</div>
-
-		<a
-			href={profileURL}
-			target="_blank"
-			class="link-focus mx-auto mt-6 flex w-max flex-row items-center justify-center gap-2"
-		>
-			<div class="h-10 w-10 overflow-hidden rounded-full bg-black bg-zinc-200 dark:bg-zinc-800">
-				<Image
-					src={image}
-					alt={name}
-					class="h-full w-full object-cover"
-					quality={100}
-					loading="lazy"
-				/>
-			</div>
-
-			<div class="mt-auto flex flex-col items-start justify-center">
-				<span
-					class="text-base font-medium text-zinc-700 transition-[text-color] duration-300 ease-in group-hover:text-zinc-900 dark:text-zinc-300 dark:group-hover:text-zinc-100 lg:text-lg"
-					>{name}</span
-				>
-				<span
-					class="ease-ingroup-hover:text-zinc-700 text-sm text-zinc-500 transition-[text-color] duration-300 dark:text-zinc-400 dark:group-hover:text-zinc-300 lg:text-base"
-					>{role}, {company}</span
-				>
-			</div>
-		</a>
-	</div>
-</article>
