--- src/modules/landing/TestimonialGroup.astro
@@ -1,12 +1,0 @@
----
-import TestimonialCard from './TestimonialCard.astro';
-import { testimonials } from '@consts';
----
-
-<div class="grid grid-cols-1 gap-4 md:grid-flow-row-dense md:grid-cols-6">
-	{
-		testimonials.map((testimonial, index) => (
-			<TestimonialCard {...testimonial} center={index === testimonials.length - 1} />
-		))
-	}
-</div>
