+++ src/shared/components/SocialLinkGroup.astro
@@ -1,0 +1,19 @@
+---
+import SocialLink from '@shared/components/SocialLink.astro';
+import {
+	LINKEDIN_PROFILE_URL,
+	GITHUB_PROFILE_URL,
+	X_PROFILE_URL,
+	BLOG_URL,
+} from '@config/consts/urls';
+---
+
+<div class="flex items-center gap-x-2 text-sm md:gap-x-6">
+	<SocialLink href={LINKEDIN_PROFILE_URL} label="LinkedIn" iconName="linked-in" />
+	<span class="hidden text-zinc-400 dark:text-zinc-500 md:block">|</span>
+	<SocialLink href={GITHUB_PROFILE_URL} label="GitHub" iconName="github" />
+	<span class="hidden text-zinc-400 dark:text-zinc-500 md:block">|</span>
+	<SocialLink href={X_PROFILE_URL} label="X" iconName="x" />
+	<span class="hidden text-zinc-400 dark:text-zinc-500 md:block">|</span>
+	<SocialLink href={BLOG_URL} label="Blog" iconName="book" />
+</div>
