+++ src/shared/layouts/Layout.astro
@@ -1,0 +1,45 @@
+---
+import type { ComponentProps } from 'astro/types';
+
+import SEO from '@shared/components/SEO.astro';
+import Header from '@shared/layouts/Header.astro';
+import Footer from '@shared/layouts/Footer.astro';
+import { DEFAULT_SEO_TITLE, DEFAULT_SEO_DESCRIPTION } from '@config/consts';
+
+type Props = ComponentProps<typeof SEO> & {
+	showFooter?: boolean;
+};
+
+const DEFAULT_TITLE = DEFAULT_SEO_TITLE;
+const DEFAULT_DESCRIPTION = DEFAULT_SEO_DESCRIPTION;
+
+const {
+	title = DEFAULT_TITLE,
+	description = DEFAULT_DESCRIPTION,
+	showFooter = true,
+	url,
+} = Astro.props;
+---
+
+<html
+	style={{
+		scrollBehavior: 'smooth',
+	}}
+	lang="en"
+>
+	<SEO {title} {description} {url} />
+	<body
+		style={{
+			scrollBehavior: 'smooth',
+		}}
+		class="bg-white dark:bg-zinc-950 dark:text-zinc-50"
+	>
+		<div class="flex flex-col">
+			<Header />
+			<main id="main">
+				<slot />
+			</main>
+		</div>
+		{showFooter ? <Footer /> : null}
+	</body>
+</html>
