+++ src/shared/pages/home/SectionHeading.astro
