+++ src/shared/types/index.ts
@@ -1,0 +1,3 @@
+import { Themes } from '@/config/consts';
+
+export type Theme = (typeof Themes)[keyof typeof Themes];
